# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "WiGO PRO — API mobile",
        "version": "1.0.0",
        "description": "Contrat entre le back-office Laravel et l'application mobile **WiGO PRO**\n(chauffeurs Yango en Côte d'Ivoire, opéré par At Confort Plus).\n\n- Montants en **entiers FCFA** (XOF).\n- Dates au format **ISO 8601 UTC**.\n- Identifiants **ULID**.\n- Messages d'erreur **en français**, prêts à afficher.\n- Enveloppe unique : succès `{ message, data }`, erreur\n  `{ message, errors }`. Le code HTTP porte le statut, jamais le corps.\n- Pagination **curseur** : `meta.next_cursor` et `links.next`,\n  `per_page` ≤ 50.\n- Authentification par **jeton Sanctum** (habilitation `mobile:*`) obtenu\n  via `POST /auth/otp/verify`.\n\nMessagerie en direct (WebSocket) : voir [`/docs/api/guides/realtime`](/docs/api/guides/realtime)."
    },
    "servers": [
        {
            "url": "{origin}/api/v1",
            "description": "Serveur courant",
            "variables": {
                "origin": {
                    "default": "http://localhost:8000",
                    "description": "Origine de l'instance (schéma, domaine, port)."
                }
            }
        }
    ],
    "security": [
        {
            "bearer": []
        }
    ],
...
    "paths": {
        "/auth/otp/request": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "operationId": "v1.auth.otp.request",
                "description": "Envoie un code à 6 chiffres (valable 5 minutes) par SMS ou WhatsApp. Le\nconducteur doit être pré-enregistré par le back-office : l'application\nmobile ne crée pas de compte.\n\nLimité à 3 envois par tranche de 10 minutes et par numéro.\n\nEn développement et en test (`WIGO_OTP_EXPOSE_CODE`), la réponse contient\nen plus le champ `code` : il est toujours absent en production.",
                "summary": "Demander un code OTP",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OtpRequestRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "examples": [
                                                "Un code de connexion vous a été envoyé. Il expire dans 5 minutes."
                                            ]
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "channel": {
                                                    "type": "string",
                                                    "enum": [
                                                        "sms",
                                                        "whatsapp"
                                                    ]
                                                },
                                                "expires_at": {
                                                    "type": "string",
                                                    "examples": [
                                                        "2026-08-29T09:47:11Z"
                                                    ]
                                                },
                                                "code": {
                                                    "type": "string",
                                                    "description": "Uniquement en développement et en recette\n(`WIGO_OTP_EXPOSE_CODE`) : absent en production.",
                                                    "examples": [
                                                        "482913"
                                                    ]
                                                }
                                            },
                                            "required": [
                                                "channel",
                                                "expires_at"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "message",
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                },
                "security": []
            }
        },
...
        "/broadcasting/auth": {
            "get": {
                "tags": [
                    "Broadcast"
                ],
                "operationId": "v1.broadcasting.auth.show",
                "summary": "Autoriser l'abonnement à un canal privé (GET)",
                "description": "Variante GET de l'opération ci-dessous, acceptée par la route du\nframework. Les SDK Pusher postent : préférer `POST`, documenté plus bas.",
                "parameters": [
                    {
                        "name": "socket_id",
                        "in": "query",
                        "required": true,
                        "description": "Identifiant de la connexion WebSocket.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "channel_name",
                        "in": "query",
                        "required": true,
                        "description": "Nom complet du canal, préfixe `private-` inclus.",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Abonnement autorisé : la signature à renvoyer au serveur WebSocket.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "auth": {
                                            "type": "string",
                                            "description": "Signature `clé:hmac` attendue par Reverb.",
                                            "examples": [
                                                "app-key:9f1c2b7e4d3a4f8b9c2e1a5d6b7c8e90"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "auth"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "403": {
                        "description": "Le canal n'appartient pas au conducteur authentifié, ou n'existe pas."
                    }
                }
            },
            "post": {
                "tags": [
                    "Broadcast"
                ],
                "operationId": "v1.broadcasting.auth",
                "summary": "Autoriser l'abonnement à un canal privé",
                "description": "Appelée automatiquement par le client Pusher avant de souscrire à un\ncanal privé ; le client y présente le jeton Sanctum du conducteur.\n\nUn canal inconnu, ou qui n'appartient pas au conducteur authentifié,\nrépond **403** — jamais 404, pour ne rien révéler de l'existence d'une\nressource à qui n'y a pas droit.\n\nNe pas confondre avec `/broadcasting/auth` (sans le préfixe `api/v1`),\nqui attend une session de navigateur et refusera un jeton Sanctum.\n\nLes règles d'accès aux canaux vivent dans `routes/channels.php`. Voir le\nguide [« Temps réel »](/docs/api/guides/realtime).",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "socket_id": {
                                        "type": "string",
                                        "description": "Identifiant de la connexion WebSocket, fourni par le client Pusher.",
                                        "example": "123456.789012"
                                    },
                                    "channel_name": {
                                        "type": "string",
                                        "description": "Nom complet du canal, préfixe `private-` inclus.",
                                        "example": "private-conversation.01HZY3M4N5P6Q7R8S9T0V1W2X3"
                                    }
                                },
                                "required": [
                                    "socket_id",
                                    "channel_name"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Abonnement autorisé : la signature à renvoyer au serveur WebSocket.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "auth": {
                                            "type": "string",
                                            "description": "Signature `clé:hmac` attendue par Reverb.",
                                            "example": "app-key:9f1c2b7e4d3a4f8b9c2e1a5d6b7c8e90"
                                        }
                                    },
                                    "required": [
                                        "auth"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "403": {
                        "description": "Le canal n'appartient pas au conducteur authentifié, ou n'existe pas."
                    }
                }
            }
        }
    },
    "components": {
...
        "responses": {
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview.",
                                    "examples": [
                                        "Authentification requise."
                                    ]
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
...
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview.",
                                    "examples": [
                                        "Les données envoyées sont invalides."
                                    ]
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "examples": [
                                                "Le champ téléphone doit être un numéro de téléphone valide."
                                            ]
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            }
        },
        "schemas": {
...
            "OtpChannel": {
                "type": "string",
                "enum": [
                    "whatsapp",
                    "sms"
                ],
                "title": "OtpChannel"
            },
            "OtpRequestRequest": {
                "type": "object",
                "properties": {
                    "phone": {
                        "type": "string",
                        "pattern": "^\\+[1-9]\\d{7,14}$",
                        "examples": [
                            "+2250717738299"
                        ]
                    },
                    "channel": {
                        "$ref": "#/components/schemas/OtpChannel"
                    }
                },
                "required": [
                    "phone"
                ],
                "title": "OtpRequestRequest"
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}